requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
bottle==0.12.25
bottle-websocket==0.2.9
Brotli==1.1.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
Eel==0.16.0
frozenlist==1.4.1
future==1.0.0
g4f==0.2.6.2
gevent==24.2.1
gevent-websocket==0.10.1
google-api-core==2.18.0
google-api-python-client==2.123.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.63.0
greenlet==3.0.3
httplib2==0.22.0
idna==3.6
multidict==6.0.5
oauthlib==3.2.2
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
uritemplate==4.1.1
urllib3==2.2.1
whichcraft==0.6.1
yarl==1.9.4
zope.event==5.0
zope.interface==6.2
